+++ daiquiri_core/templates/core/base_navigation.html
@@ -41,7 +41,7 @@
                                 <a href="{% url 'messages' %}">{% trans 'Contact messages' %}</a>
                             </li>
                             <li>
-                                <a href="{% url 'metadata' %}">{% trans 'Metadata management' %}</a>
+                                <a href="{% url 'metadata_management' %}">{% trans 'Metadata management' %}</a>
                             </li>
                             <li>
                                 <a href="{% url 'examples' %}">{% trans 'Query examples' %}</a>
