+++ daiquiri_metadata/templates/metadata/management_modal_delete_functions.html
