+++ daiquiri_metadata/templates/metadata/management_options.html
