--- daiquiri_metadata/templates/metadata/metadata_database_view.html
@@ -1,21 +1,0 @@
-{% extends 'core/page.html' %}
-{% load i18n %}
-
-{% block page %}
-
-    <h1> {{ db_name }}</h1>
-
-         <h2> {% trans 'Description' %}</h2>
-            <p> {{ db_description }} </p>
-
-         <h2> {% trans 'List of tables' %} </h2>
-            <ul class="list-unstyled">
-                {% for table in db_tables %}
-                <li>
-                    <a href="{{ table.name }}"> {{ table }} </a>
-                    <p> {{ table.description }}
-                </li>
-                {% endfor %}
-            </ul>
-
-{% endblock %}
